requirements.txt
imutils==0.5.4
joblib==1.1.0
numpy==1.21.4
opencv-contrib-python==4.5.4.58
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
PyQt5 >= 5.12
PyQt5-sip
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0

